Task: Update protocol of book cover URL to HTTPS
When deployed in production (https), setting the image source to a URL over http
will send errors to the browser console about insecure contexts.

This change tests if the deployement is production and if so replaces instances
of 'http:' with 'https:'.

Development deployments default to HTTP so no changes are required.